Remove outbound resolver endpoints.
There's no reason to put a resolver endpoint in each VPC.  Resolver rule
are associated with the *resolver service* in each VPC.  That's all a
black box.

A resolver rule basically intercepts an incoming DNS request and
"relocates it" to an outbound resolver endpoint.  It's imnportant to
note that these resolver endpoints are rate-limited to 10,000 queries
per second per IP.

Cutting down on resolver endpoints significantly reduces the costs
of resolver services.